pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel>=0.40.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fishontario"
version = "1.0.0"
description = "Ontario Fishing Regulations QA API"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.115.9",
    "uvicorn==0.23.2",
    "python-dotenv==1.0.0",
    "langchain-core==0.3.60",
    "langchain-openai==0.0.5",
    "openai==1.6.1",
    "pypdf==3.17.1",
    "tiktoken==0.5.2",
    "pydantic==2.11.5",
    "python-multipart==0.0.6",
    "numpy>=1.24.3,<1.25.0",
    "typing-extensions>=4.5.0",
    "requests>=2.31.0"
] 